Fix behavior of SeekPrefix()
With 0ba5646453fb1e36d72bed648463fdd9f4b15e74 we now added a <prefix>
key into the database. This key is used for optimziation purposes and
should be ignored. SeekPrefix() had a bug where it would seek to
<prefix> when target was empty. This means that any SeekPrefix("") would
end up at <prefix> and report as !Valid(). This patches fixes it