Refresh iterator snapshots on commit/abort
Once the transactions commits or aborts, we need to refresh all live
iterator's snapshot. I assumed that higher level code would yield all of
their iterators on commit, but this seems not to be the case (at least it's not
true when building indexes in the background). For that reason, we now
keep track of all live iterators and manually refresh their snapshots on
each commit/abort.

This patch also adds transaction control to RocksStandardIndex, which
fixes the concurrency when building index in the background.