SERVER-19703 Bump the RocksDB format version
This commit makes it so that MongoRocks databases created with 3.4
version cannot be opened by 3.2 version. This is important because
MongoRocks 3.4 creates index in a new format. However, we also retain
backwards compatibility -- if a database was created with MongoRocks 3.2
the new version will continue working in compatibility mode and keep
creating new indexes in the old format.