SERVER-27497 Do not rely on obsolete modifications to the global environment (#51)
SConscripts should work on a local copy of the environment. This has been changed in the upstream mongo sources, and a side effect is that some previously (unintentional) global changes (like making the mongo headers available), are now local. The rocks repo needs to now explicitly modify the environment. In order to avoid mutating the global environment, a local Clone should be made.