use fakeCompator to skip oplog cf trim operation
The rollback logic at the mongo-server level calls the rtt(rollback to timestamp) function of the storage engine layer.
This requires user data rollback, but oplog cannot be rolled back. However, at present, the rollback operation of rocksdb is done at the opendb time,
and no parameters can be set to decide which cf needs to be rolled back and which cf cannot be rolled back.
Inside rocksdb, it is only based on the timestamp turned on at the cf time to decide whether to roll back.
Therefore, in order to meet the oplog non-roll policy required by the upper level of mongo,
we temporarily set oplog cf not support ts to make rocksdb skip the rollback behavior of oplog cf.