remove _db, _oplogManger, _snapshotManger, _compactionSCheduler, _durabilityManager from rocks recover unit
Because rtt cleans historical data in the reopen db way,
there will not be a stable db pointer at the entire mongo-server level.
Because it needs to be reopened, the db pointer has also changed.
Given this logic, we removed the _db, _oplogManger, _snapshotManger, _compactionSCheduler, _durabilityManager
these in the form of naked pointers in the rocks recover unit
and instead fetch them from the rocks engine each time.

reorder compactionScheduler init before initdatabase